backend/requirements.txt
# ── Core Django stack ───────────────────────────────────────
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0

# ── Upgrade 1: Placement Predictor ──────────────────────────
scikit-learn>=1.3
numpy>=1.24

# ── Upgrade 2: Resume Parser ────────────────────────────────
pdfplumber>=0.10

# ── Upgrade 3: Sentiment Analysis ───────────────────────────
textblob>=0.17

# ── Upgrade 4: Celery + Redis ────────────────────────────────
celery>=5.3
redis>=5.0
kombu>=5.3

# ── Upgrade 5: Django Channels ──────────────────────────────
channels>=4.0
channels-redis>=4.1
daphne>=4.0

# ── Upgrade 6: Report Exports ────────────────────────────────
reportlab>=4.0
openpyxl>=3.1

# ── Groq AI Integration ──────────────────────────────────────
groq>=0.9.0

# ── Production deployment ────────────────────────────────────
gunicorn>=21.2          # WSGI server for production
whitenoise>=6.6         # Serve static files without a CDN
psycopg2-binary>=2.9    # PostgreSQL driver
dj-database-url>=2.1    # Parse DATABASE_URL env variable
python-dotenv>=1.0      # Load .env files
